Factored out threading model from TcpIpConnectionManager
