+++ hazelcast-client-new/src/main/java/com/hazelcast/client/connection/nio/AbstractClientSelectionHandler.java
@@ -19,8 +19,8 @@ package com.hazelcast.client.connection.nio;
 import com.hazelcast.client.connection.ClientConnectionManager;
 import com.hazelcast.logging.ILogger;
 import com.hazelcast.logging.Logger;
-import com.hazelcast.nio.tcp.IOSelector;
-import com.hazelcast.nio.tcp.SelectionHandler;
+import com.hazelcast.nio.tcp.nonblocking.IOSelector;
+import com.hazelcast.nio.tcp.nonblocking.SelectionHandler;
 import com.hazelcast.nio.tcp.SocketChannelWrapper;
 
 import java.nio.channels.SelectionKey;
