+++ hazelcast/src/main/java/com/hazelcast/nio/tcp/iobalancer/EventCountBasicMigrationStrategy.java
@@ -16,7 +16,7 @@
 
 package com.hazelcast.nio.tcp.iobalancer;
 
-import com.hazelcast.nio.tcp.MigratableHandler;
+import com.hazelcast.nio.tcp.nonblocking.MigratableHandler;
 import java.util.Set;
 
 /**
