+++ hazelcast/src/main/java/com/hazelcast/nio/tcp/iobalancer/LoadTracker.java
@@ -17,9 +17,9 @@
 package com.hazelcast.nio.tcp.iobalancer;
 
 import com.hazelcast.logging.ILogger;
-import com.hazelcast.nio.tcp.AbstractIOSelector;
-import com.hazelcast.nio.tcp.IOSelector;
-import com.hazelcast.nio.tcp.MigratableHandler;
+import com.hazelcast.nio.tcp.nonblocking.AbstractIOSelector;
+import com.hazelcast.nio.tcp.nonblocking.IOSelector;
+import com.hazelcast.nio.tcp.nonblocking.MigratableHandler;
 import com.hazelcast.util.ItemCounter;
 
 import java.util.HashMap;
