+++ hazelcast/src/main/java/com/hazelcast/nio/tcp/iobalancer/MigrationStrategy.java
@@ -16,7 +16,8 @@
 
 package com.hazelcast.nio.tcp.iobalancer;
 
-import com.hazelcast.nio.tcp.MigratableHandler;
+import com.hazelcast.nio.tcp.nonblocking.MigratableHandler;
+import com.hazelcast.nio.tcp.nonblocking.IOSelector;
 
 /**
  * Decides if a {@link MigratableHandler handler} migration should be attempted
@@ -27,7 +28,7 @@ import com.hazelcast.nio.tcp.MigratableHandler;
 interface MigrationStrategy {
 
     /**
-     * Looks for imbalance in {@link MigratableHandler handler} to {@link com.hazelcast.nio.tcp.IOSelector ioSelector}
+     * Looks for imbalance in {@link MigratableHandler handler} to {@link IOSelector ioSelector}
      * mapping.
      *
      * @param imbalance
