+++ hazelcast/src/main/java/com/hazelcast/nio/tcp/nonblocking/IOSelectorOutOfMemoryHandler.java
@@ -14,18 +14,9 @@
  * limitations under the License.
  */
 
-package com.hazelcast.nio.tcp;
+package com.hazelcast.nio.tcp.nonblocking;
 
-import com.hazelcast.client.impl.protocol.ClientMessage;
+public interface IOSelectorOutOfMemoryHandler {
 
-import java.nio.ByteBuffer;
-
-public class SocketClientMessageWriter
-        implements SocketWriter<ClientMessage> {
-
-    @Override
-    public boolean write(ClientMessage clientMessage, ByteBuffer socketBuffer)
-            throws Exception {
-        return clientMessage.writeTo(socketBuffer);
-    }
+    void handle(OutOfMemoryError error);
 }
